Don't build if we aren't on master

+++ build_helpers/buildDistFromTravis.sh
@@ -1,17 +1,24 @@
 #!/bin/bash
 set -e
 
-# build dist
-npm run build-dist
+branch_name="$(git symbolic-ref HEAD 2>/dev/null)"
 
-# initialize and commit everything in pages
-cd dist
-git init
-git config user.name "Travis"
-git config user.email "Travis"
-git add .
-git commit -m "build dist assets and commit to dist"
+if [ "$branch_name" == "refs/heads/master" ]
+then
+  # build dist
+  npm run build-dist
 
-# hide my output for security reasons
-# force it because we want this to just actually work
-git push --force --quiet "https://${GH_TOKEN}@${GH_REF}" master:dist > /dev/null 2>&1
+  # initialize and commit everything in pages
+  cd dist
+  git init
+  git config user.name "Travis"
+  git config user.email "Travis"
+  git add .
+  git commit -m "build dist assets and commit to dist"
+
+  # hide my output for security reasons
+  # force it because we want this to just actually work
+  git push --force --quiet "https://${GH_TOKEN}@${GH_REF}" master:dist > /dev/null 2>&1
+else
+  echo "Current branch is not master and dist will not be built or committed."
+fi
